backend/pyproject.toml
[tool.poetry]
name = "todo-backend-api"
version = "0.1.0"
description = "Todo Backend API with FastAPI, SQLModel, and JWT authentication"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
black = "^24.0.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true
app_dir = "src"